contract clauses detector/icis/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
